TbKafkaSettingsTest instead of multiples
